Amit ma kijavtottunk.

+++ src/quixo/heuristics/Heuristics.java
@@ -5,7 +5,7 @@ import quixo.players.minimax.Node;
 public interface Heuristics {
 	void init(int me, int you, int nobody);
 	void empty();
-	int sum();
-	int calculation(Node node);
-	int getValue();
+	double sum();
+	double calculation(Node node);
+	double getValue();
 }
