+++ src/quixo/heuristics/Index.java
@@ -10,7 +10,7 @@ public class Index extends SimpleHeuristic{
 		super.init( me, you, nobody);
 	}
 	
-	public int calculation(Node node){
+	public double calculation(Node node){
 		this.node=node;
 		table=(QuixoBoard) node.getTable().clone();
 		color=(node.getModel()+1)%2;
